# docs/openapi.generated.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Met à jour le profil",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/interests": {
      "put": {
        "operationId": "UsersController_updateInterests",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInterestsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Met à jour les centres d’intérêt",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/streak": {
      "get": {
        "operationId": "UsersController_streak",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Streak de lecture",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/push-tokens": {
      "post": {
        "operationId": "UsersController_registerPush",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPushTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enregistre un token push Expo",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/push-tokens/{token}": {
      "delete": {
        "operationId": "UsersController_removePush",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Désenregistre un token push",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des membres (staff)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/new": {
      "get": {
        "operationId": "UsersController_newMembers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Nouveaux membres (signaux People)",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/sermons": {
      "get": {
        "operationId": "SermonsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "series",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "live",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des sermons",
        "tags": [
          "sermons"
        ]
      },
      "post": {
        "operationId": "SermonsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSermonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crée / planifie un sermon (staff)",
        "tags": [
          "sermons"
        ]
      }
    },
    "/api/v1/sermons/{id}": {
      "get": {
        "operationId": "SermonsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail sermon",
        "tags": [
          "sermons"
        ]
      },
      "patch": {
        "operationId": "SermonsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSermonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Met à jour un sermon (staff)",
        "tags": [
          "sermons"
        ]
      }
    },
...
  },
  "info": {
    "title": "IPCK API",
    "description": "API de la plateforme IPCK (mobile + dashboard)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        }
      },
      "UpdateInterestsDto": {
        "type": "object",
        "properties": {
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RegisterPushTokenDto": {
        "type": "object",
        "properties": {
          "expoToken": {
            "type": "string",
            "example": "ExponentPushToken[xxxxxxxx]"
          },
          "platform": {
            "type": "string",
            "example": "android"
          }
        }
      },
...
      "CreateSermonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "speaker": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "series": {
            "type": "string"
          },
          "live": {
            "type": "boolean"
          },
          "videoUrl": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "published"
            ]
          }
        },
        "required": [
          "title",
          "speaker",
          "date",
          "duration",
          "series"
        ]
      },
      "UpdateSermonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "speaker": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "series": {
            "type": "string"
          },
          "live": {
            "type": "boolean"
          },
          "videoUrl": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "published"
            ]
          }
        }
      },
...
    }
  }
}